Got rid of sign-out error

+++ lib/actions/user.actions.ts
@@ -46,7 +46,9 @@ export async function signInWithCredentials(
 export async function signOutUser() {
   // get current users cart and delete it so it does not persist to next user
   const currentCart = await getMyCart();
-  await prisma.cart.delete({ where: { id: currentCart?.id } });
+  if (currentCart?.userId) {
+    await prisma.cart.delete({ where: { id: currentCart?.id } });
+  }
   await signOut();
 }
